LPD-84021 Track the row count with a local long counter and assert the enablement warning.

+++ modules/apps/portal/portal-upgrade-impl/src/test/java/com/liferay/portal/upgrade/internal/report/UpgradeReportTest.java
@@ -321,7 +321,7 @@ public class UpgradeReportTest {
 			Mockito.mock(Connection.class)
 		);
 
-		int currentRow = RandomTestUtil.randomInt();
+		long currentRow = RandomTestUtil.randomLong();
 
 		String upgradeProcessClassName =
 			"com.liferay.test.SampleUpgradeProcess";
